+++ NexmoClient/Internal/NexmoCore/Network/Request/NXMInviteUserRequest.h
@@ -11,12 +11,12 @@
 @interface NXMInviteUserRequest : NXMBaseRequest
 
 @property (nonatomic, strong, nonnull) NSString *conversationID;
-@property (nonatomic, strong, nonnull) NSString *userID;
+@property (nonatomic, strong, nonnull) NSString *username;
 @property (nonatomic) BOOL mediaEnabled;
 
-- (nullable instancetype)initWithConversationId:(nonnull NSString *)conversationID andUserID:(nonnull NSString *)userID;
+- (nullable instancetype)initWithConversationId:(nonnull NSString *)conversationID username:(nonnull NSString *)username;
 
-- (nullable instancetype)initWithConversationId:(nonnull NSString *)conversationID andUserID:(nonnull NSString *)userID mediaEnabled:(BOOL)mediaEnabled;
+- (nullable instancetype)initWithConversationId:(nonnull NSString *)conversationID username:(nonnull NSString *)username mediaEnabled:(BOOL)mediaEnabled;
 
 @end